% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_any.R
\name{any}
\alias{any}
\title{Any -- a "Fits-All" Type}
\description{
This type does not accept any additional arguments passed as \code{...} to
\code{\link{declare}} and \code{\link{cast}}.
}
\details{
Any R object is a valid instance of this type.
}
\concept{types}
