% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{g}
\alias{g}
\title{Get Value of a Typed Variable}
\usage{
g(x, env = parent.frame(), inherits = FALSE, .character = FALSE)
}
\arguments{
\item{x}{the variable of which the value should be returned.}

\item{env}{the environment in which to look for \code{x}. Defaults to the
current environment.}

\item{inherits}{a logical value indicating if the enclosing frames of the
environment should be inspected.}

\item{.character}{a logical value indicating if \code{x} is the variable to
get or a character string with the name of the variable to get.}
}
\value{
x
}
\description{
Return the value of a typed variable if it is valid, otherwise throw an
error.
}
\seealso{
\link{typed}, \code{\link{\%<-\%}} and \code{\link{\%<-s\%}} for
  securely assigning to a typed variable
}

