% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnvOfObject.R
\name{getEnvOfObject}
\alias{getEnvOfObject}
\title{Get the Environment Where an Object is Defined}
\usage{
getEnvOfObject(x, env)
}
\arguments{
\item{x}{the name of the object to find as a character string.}

\item{env}{the environment where to start the search.}
}
\value{
the first environment along the search path where an object with the
  name \code{x} is defined, or the empty environment if the search is not
  successful.
}
\description{
This recursive function returns the environment where the object with a
particular name is defined, starting with the search in a given environment.
It used R's standard scoping rules, with the exception that it does not
recurse beyond the global environment into the packages' namespaces.
}
\keyword{internal}

