% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLHSVariableName.R
\name{getLHSVariableName}
\alias{getLHSVariableName}
\title{Get Variable Name From Left-Hand Side of Assignment}
\usage{
getLHSVariableName(x)
}
\arguments{
\item{x}{the left-hand side of an assignment as a language object.}
}
\value{
a character string.
}
\description{
\code{getLHSVariableName} is a recursive function which returns the name of
the variable which would be created or modified by a \emph{valid} assignment.
}
\examples{
assignment <- quote(names(var)[1:3] <- letters[1:3])
lhs <- assignment[[2]]  # left-hand side of assignment
getLHSVariableName(lhs)  # "var"
}
\keyword{internal}

