% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_c.R
\name{iMode}
\alias{iMode}
\title{Enhanced Mode/Type Determination}
\usage{
iMode(x)
}
\arguments{
\item{x}{any R object.}
}
\value{
\describe{
  \item{"ordered"}{for ordered factors}
  \item{"factor"}{for (unordered) factors}
  \item{"data.frame"}{for lists which inherit from "data.frame"}
  \item{"\emph{mode}Matrix"}{for matrices, where \emph{mode} is
                             \code{mode(x)}}
  \item{"\emph{mode}Array"}{for (other) arrays, where \emph{mode} is
                            \code{mode(x)}}
  \item{"namespace"}{for namespaces}
  \item{typeof(x)}{for everything else; types "integer" and "double" are
                   returned as "numeric" and types "special" and "builtin"
                   are returned as "function"}
}
}
\description{
Get the high-level mode/type of an object.
}

