% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkType.R
\name{markInvalidWError}
\alias{markInvalidWError}
\title{Mark the Variable Currently Under Validation as Invalid}
\usage{
markInvalidWError(...)
}
\arguments{
\item{...}{one or more character strings to be pasted together to an error
message (without a separator), which describes why a variable is not of the
declared type.}
}
\description{
Mark the Variable Currently Under Validation as Invalid
}
\keyword{internal}

