% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_numeric.R
\name{numeric}
\alias{numeric}
\title{Typed Numeric Vectors}
\arguments{
\item{length}{the exact length of the vector.}

\item{min_length, max_length}{the minimum/maximum length of the vector
(ignored if \code{lenght} is given).}

\item{*set}{the set of allowed values as a numeric vector.}

\item{min, max}{the minimum/maximum values allowed (ignored if \code{set} is
given).}

\item{allow_NA, allow_NaN}{are \code{NA} and \code{NaN} values allowed?
\code{FALSE} for \code{allow_NA} or \code{allow_NaN} is ignored here if
\code{set} contains \code{NA} or \code{NaN} respectively.}
}
\description{
The following arguments can be passed as \code{...} to \code{\link{declare}}
and \code{\link{cast}}.
}
\details{
Properties marked with a * are checked using non-primitive
  functions, which increases the reference count of the object. Hence, after
  checking the validity of a variable with such a property, it can no longer
  be modified in place, i.e the next modification of it will result in a
  copy of it being made.
}
\seealso{
Other types: \code{\link{any}}, \code{\link{character}}
}

