% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerType.R
\name{registerCustomType}
\alias{registerCustomType}
\title{Register Custom Types}
\usage{
registerCustomType(type, checkPropertiesFun, check_type_expr,
  varname = ".XXX.")
}
\arguments{
\item{type}{a character string with the name of the type to register.}

\item{checkPropertiesFun}{a function to check the validity of the additional
properties passed to either declare or cast when creating a typed variable.
This function must accept these additional properties as (named) parameters
and it should throw an error if properties used are invalid. Afterwards, it
must return -- as a (named) list -- the set of properties which should
finally be stored. This set can be different from the original arguments to
this function, e.g. because one of two inconsistent properties was removed.
There is the helper function \code{\link{args2list}}, which returns the
*current* values of the arguments of the calling function as a list.}

\item{check_type_expr}{a quoted expression to be used for checking whether a
 variable is of type \code{type}. Within this expression, refer to the
 variable to check by what is specified in the \code{varname}-argument (as a
 name, i.e. without the quotes). The default is \code{.XXX.}. The list of
 additional properties is available as \code{.lazyTyper_properties}.

 This expression should never fail, i.e. throw an error. To indicate that the
 variable under consideration is not of type \code{type}, the logical
 variable \code{.lazyTyper_valid} should be set to \code{FALSE}. In addition,
 the attribute "error" of \code{.lazyTyper_valid} (a character vector),
 should describe the reason(s) why the variable is invalid. The preferred way
 to both set \code{.lazyTyper_valid} to \code{FALSE} and to add an
 (additional) message to its attribute "error", is to call
 \code{markInvalidWError}.}

\item{varname}{a character string with the name of the placeholder-variable
used in \code{check_type_expr} to be replaced with the actual variable to
validate.}
}
\description{
Register Custom Types
}
\seealso{
useful helper functions: \code{\link{markInvalidWError}},
 \code{\link{args2list}}, \code{\link{hasValue}}
}

