% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove}
\alias{remove}
\alias{rm}
\title{A Replacement for base \code{\link[base]{remove}}/\code{\link[base]{rm}} with
Support for Typed Variables}
\usage{
remove(..., list = character(), pos = -1, envir = as.environment(pos),
  inherits = FALSE, .untype = TRUE)

rm(..., list = character(), pos = -1, envir = as.environment(pos),
  inherits = FALSE, .untype = TRUE)
}
\arguments{
\item{...}{the objects to be removed, as names (unquoted) or character
strings (quoted).}

\item{list}{a character vector naming objects to be removed.}

\item{pos}{where to do the removal. By default, uses the current environment.
See ‘Details’ in the documentation of the base functions for other
possibilities.}

\item{envir}{the environment to use. See ‘Details’ in the documentation of
the base functions for other possibilities.}

\item{inherits}{should the enclosing frames of the environment be inspected?}

\item{.untype}{should typed variables be untyped?}
}
\description{
These functions differ from the base versions only in that they (optionally)
untype typed variables.
}
\details{
See `Details` in base
  \code{\link[base]{remove}}/\code{\link[base]{rm}}.
}
\seealso{
\link{typed}
}
