% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scope.R
\docType{data}
\name{SCOPE}
\alias{SCOPE}
\alias{\%->\%}
\title{Scopes for Limited Variable Lifetime}
\format{An object of class \code{character} of length 1.}
\usage{
SCOPE \%->\% expr

SCOPE
}
\arguments{
\item{SCOPE}{an object with the name "SCOPE". It is not evaluated and only
its name matters. Such a dummy object is exported by this package to enable
proper syntax highlighting and code diagnostics in editors/IDEs like
RStudio.}

\item{expr}{an expression to evaluate in the calling environment of
\code{\%->\%}. Variables \code{\link{declare}}d \emph{directly} herein are
untyped and deleted at the end of the scope.}
}
\value{
whatever \code{expr} returns.
}
\description{
Variables which are \code{\link{declare}}d in a scope are untyped and deleted
at the end of it.
}
\note{
Scopes can be nested.
}
\examples{
declare(a, "numeric")
SCOPE \%->\% {
  declare(b, "numeric")
  b <- 2
  a \%<-\% .(b * 3)
  d <- 99
}
a
d
exists("b")
is.typed(b)
}
\keyword{datasets}

