% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_c.R
\name{hasValue}
\alias{hasValue}
\title{Does an Argument (or Another Object) Exist and have a Value?}
\usage{
hasValue(x, env = parent.frame())
}
\arguments{
\item{x}{the name of the object to test as a character string.}

\item{env}{the environment of the object to test. Defaults to the calling
frame of this function.}
}
\description{
This function returns \code{FALSE} if an object with the name \code{x} does
not exist or if it is identical to the empty name, which is used to represent
\emph{missing arguments} to a function \emph{without a default value}.
Otherwise it returns \code{TRUE}.
}
\note{
If \code{x} is a promise \code{TRUE} is returned independently of the
  value of the promise.
}
\seealso{
\code{\link[base]{exists}}, which returns \code{TRUE} also for
  missing arguments and \code{\link[base]{missing}}, which returns
  \code{TRUE} also for missing arguments which have a default value.
}
\keyword{internal}

