% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkType.R, R/registerType.R
\name{markInvalidWError}
\alias{markInvalidWError}
\alias{registerCustomType}
\title{Register Custom Types}
\usage{
markInvalidWError(...)

registerCustomType(type, checkPropertiesFun, check_type_expr,
  varname = ".XXX.")
}
\arguments{
\item{...}{one or more character strings to be pasted together to an error
message (without a separator), which describes why a variable is not of the
declared type.}

\item{type}{a character string with the name of the type to register.}

\item{checkPropertiesFun}{a function to be used for checking whether the
additional properties supplied when typing a variable to type \code{type}
are correct (no missing compulsory properties, no unknown properties, no
inconsistent ones etc.). It must accept a single formal argument \code{x},
a potentially empty named list. It should throw an error if the supplied
list of additional properties is invalid.}

\item{check_type_expr}{a quoted expression to be used for checking whether a
  variable is of type \code{type}. Within this expression, refer to the
  variable to check by what is specified in the \code{varname}-argument (as a
  name, i.e. without the quotes). The default is \code{.XXX.}. The list of
  additional properties is available as \code{.lazyTyper_properties}.

  This expression should never fail, i.e. throw an error. To indicate that
  the variable under consideration is not of type \code{type}, the logical
  variable \code{.lazyTyper_valid} should be set to \code{FALSE}. In
  addition, the attribute "error" of \code{.lazyTyper_valid} (a character
  vector), should describe the reason(s) why the variable is invalid. The
  preferred way to both set \code{.lazyTyper_valid} to \code{FALSE} and to
  add an (additional) message to its attribute "error", is to call
  \code{markInvalidWError}.}

\item{varname}{a character string with the name of the placeholder-variable
used in \code{check_type_expr} to be replaced with the actual variable to
validate.}
}
\description{
Register Custom Types
}
\keyword{internal}

