% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerType.R
\name{registerCustomType}
\alias{registerCustomType}
\title{Register Custom Types}
\usage{
registerCustomType(type, checkPropertiesFun, checkTypeFun)
}
\arguments{
\item{type}{a character string with the name of the type to register.}

\item{checkPropertiesFun}{a function to check the validity of the additional
properties passed to either declare or cast when creating a typed variable.
This function must accept these additional properties as (named) parameters
and it should throw an error if properties used are invalid. Afterwards, it
must return -- as a (named) list -- the set of properties which should
finally be stored. This set can be different from the original arguments to
this function, e.g. because one of two inconsistent properties was removed.
There is the helper function \code{\link{args2list}}, which returns the
*current* values of the arguments of the calling function as a list.}

\item{checkTypeFun}{a function used for checking whether a variable is of type
\code{type}. This function must accept \code{x}, which is the variable to
test, as well as all the additional properties supported by this type as
(named) parameters.}
}
\description{
Register Custom Types
}
\seealso{
useful helper functions: \code{\link{markInvalidWError}},
 \code{\link{args2list}}, \code{\link{hasValue}}
}

